+++ sponsoredTransactionsAuction/backend/src/main.rs
@@ -322,8 +322,7 @@ async fn handle_signature_bid(
             used_energy: _,
         } => {
             tracing::warn!("TransactionSimulationError with reason: {:#?}.", reason);
-            let reason =
-                reason.known_or_else(|| ServerError::Unknown("RejectReason".to_string()))?;
+            let reason = reason.known_or_err()?;
 
             return Err(ServerError::TransactionSimulationError(RevertReason {
                 reason,